requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.9.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
aiosqlite>=0.19.0
httpx>=0.25.1
python-dotenv>=1.0.0
pymysql>=1.1.0
cryptography>=41.0.7
